+++ library/xgpl_src/supplementary_driver/hal/c2c/eeprom/eeprom.h
@@ -54,7 +54,7 @@ namespace HAL
    deliver the EEPROM size in uint8_t -> mult C2C BIOS value with 1024
    @return EEPROM size in byte
   */
-  inline int16_t getEepromSize(void)
+  inline uint32_t getEepromSize(void)
     {return (1024 * __HAL::get_eeprom_size());};
 
   /**
@@ -64,7 +64,7 @@ namespace HAL
   // MSCHMIDT - I think this is actually supposed to say size in bytes instead
   // MSCHMIDT - of size in kbyte.  get_eeprom_segment_size() returns size in
   // MSCHMIDT - bytes.
-  inline int16_t getEepromSegmentSize(void)
+  inline uint16_t getEepromSegmentSize(void)
     {return __HAL::get_eeprom_segment_size();};
 
   /**
