+++ library/xgpl_src/supplementary_driver/hal/esx/eeprom/eeprom.h
@@ -55,14 +55,14 @@ namespace HAL
    deliver the EEPROM size in uint8_t -> mult ESX BIOS value with 1024
    @return EEPROM size in byte
   */
-  inline int16_t getEepromSize(void)
+  inline uint32_t getEepromSize(void)
     {return (1024 * __HAL::get_eeprom_size());};
 
   /**
    deliver the EEPROM segment size in kbyte
    @return EEPROM segment size in kbyte
   */
-  inline int16_t getEepromSegmentSize(void)
+  inline uint16_t getEepromSegmentSize(void)
     {return __HAL::get_eeprom_segment_size();};
 
   /**
